Started on foundation of lexer
I want mergeable source locations because solving this problem later
means it doesn't get solved. If you are going to report where the error
is an you are at an internal AST node you need to be able to figure out
what that node corresponds to in the file. I have fought this battle
before. This is my latest attempt at something easy to use.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>